backend/api/v1/views.py
from rest_framework.viewsets import ReadOnlyModelViewSet

from recipes.models import Ingredient
from .serializers import IngredientSerializer


class IngredientViewSet(ReadOnlyModelViewSet):
    queryset = Ingredient.objects.all()
    serializer_class = IngredientSerializer
